requirements.txt
aiofiles==23.2.1
altair==5.5.0
anyio==4.2.0
argcomplete==3.1.4
attrs==23.2.0
Babel==2.10.3
bcc==0.29.1
bcrypt==3.2.2
blinker==1.7.0
Brlapi==0.8.5
cachetools==5.3.0
certifi==2023.11.17
chardet==5.2.0
click==8.1.6
cloud-init==24.4
colorama==0.4.6
command-not-found==0.3
configobj==5.0.8
contourpy==1.3.1
cryptography==41.0.7
cupshelpers==1.0
cycler==0.12.1
dbus-python==1.3.2
defer==1.0.6
distro==1.9.0
distro-info==1.7+build1
duplicity==2.1.4
fastapi==0.101.0
fasteners==0.18
fonttools==4.55.0
gensim==4.3.2
gitdb==4.0.12
GitPython==3.1.44
google-api-python-client==1.7.12
google-auth==1.5.1
google-auth-httplib2==0.2.0
greenlet==3.0.3
gunicorn==20.1.0
gyp==0.1
h11==0.14.0
httplib2==0.20.4
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.4.2
jsonpatch==1.32
jsonpointer==2.0
jsonschema==4.10.3
kiwisolver==1.4.7
language-selector==0.1
launchpadlib==1.11.0
lazr.restfulclient==0.14.6
lazr.uri==1.0.6
louis==3.29.0
Mako==1.3.2.dev0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
mdurl==0.1.2
monotonic==1.6
narwhals==1.24.0
netaddr==0.8.0
netifaces==0.11.0
nltk==3.9.1
numpy==2.1.3
oauth2client==4.1.3
oauthlib==3.2.2
olefile==0.46
packaging==24.0
pandas==2.2.3
paramiko==2.12.0
pexpect==4.9.0
pillow==10.2.0
pipx==1.4.3
platformdirs==4.2.0
protobuf==5.29.3
psutil==5.9.8
ptyprocess==0.7.0
pyarrow==19.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.25.1
pycups==2.0.1
pydantic==1.10.14
pydeck==0.9.1
PyDrive2==0.0.0
Pygments==2.17.2
PyGObject==3.48.2
PyJWT==2.7.0
PyNaCl==1.5.0
pyOpenSSL==23.2.0
pyparsing==3.1.1
pyrsistent==0.20.0
pyserial==3.5
python-apt==2.7.7+ubuntu3
python-dateutil==2.8.2
python-debian==0.1.49+ubuntu2
python-multipart==0.0.9
pytz==2024.1
pyxdg==0.28
PyYAML==6.0.1
regex==2024.11.6
requests==2.31.0
rich==13.7.1
rsa==4.9
scipy==1.15.1
seaborn==0.13.2
setuptools==68.1.2
simplejson==3.19.2
six==1.16.0
smart-open==7.1.0
smmap==5.0.2
sniffio==1.3.0
SQLAlchemy==1.4.50
starlette==0.27.0
streamlit==1.41.1
systemd-python==235
tenacity==9.0.0
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
typing_extensions==4.10.0
tzdata==2024.2
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.2
unattended-upgrades==0.1
uritemplate==4.1.1
urllib3==2.0.7
usb-creator==0.3.16
userpath==1.9.1
uvicorn==0.27.1
uvloop==0.19.0
wadllib==1.3.6
watchdog==6.0.0
wheel==0.42.0
wrapt==1.17.2
wsproto==1.2.0
xdg==5
xkit==0.0.0
